pyproject.toml
[tool.poetry]
name = "test-task"
version = "0.1.0"
description = ""
package-mode = false
authors = ["zabit <xaclafun1991@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
pydantic-settings = "^2.7.1"
environs = "^14.1.0"
uvicorn = "^0.34.0"
aiosqlite = "^0.20.0"
aiokafka = "^0.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
